final commit

+++ src/boards/boards.service.ts
@@ -5,6 +5,7 @@ import { CreateBoardDto } from './dto/create-board.dto';
 import { BoardRepository } from './board.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Board } from './board.entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class BoardsService {
@@ -15,17 +16,25 @@ export class BoardsService {
         
     ){}
 
-    async getAllBoards(): Promise<Board[]>{
-        return this.boardRepository.find();
+    async getAllBoards(
+        user: User
+    ): Promise<Board[]>{
+        const query = this.boardRepository.createQueryBuilder('board')
+
+        query.where('board.userId = :userId', { userId: user.id})
+
+        const boards = await query.getMany()
+
+        return boards;
     }
     
     // getAllBoards(): Board[] {
     //     return this.boards;
     // }
 
-    createBoard(createBoardDto: CreateBoardDto): Promise<Board>{
+    createBoard(createBoardDto: CreateBoardDto, user: User): Promise<Board>{
         
-        return this.boardRepository.createBoard(createBoardDto);
+        return this.boardRepository.createBoard(createBoardDto, user);
     }
 
     // createBoard(createBoardDto: CreateBoardDto){
@@ -60,8 +69,8 @@ export class BoardsService {
     //     }
     //     return found;
     // }
-        async deleteBoard(id: number): Promise<void>{
-            const result = await this.boardRepository.delete(id);
+        async deleteBoard(id: number, user: User): Promise<void>{
+            const result = await this.boardRepository.delete({id, userId: user.id});
 
             if(result.affected === 0) {
                 throw new NotFoundException(`Can't find Board with id ${id}`)
